+++ lib/measures/ExteriorWallThermalPropertiesPercentChange/tests/ExteriorWallThermalPropertiesMultiplier_test.rb
@@ -1,5 +1,5 @@
 # *******************************************************************************
-# OpenStudio(R), Copyright (c) 2008-2020, Alliance for Sustainable Energy, LLC.
+# OpenStudio(R), Copyright (c) 2008-2021, Alliance for Sustainable Energy, LLC.
 # All rights reserved.
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions are met:
@@ -33,5 +33,54 @@
 # OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 # *******************************************************************************
 
-require 'openstudio/calibration_measures/version'
-require 'openstudio/calibration_measures/extension'
+require 'openstudio'
+
+require 'openstudio/ruleset/ShowRunnerOutput'
+
+require "#{File.dirname(__FILE__)}/../measure.rb"
+
+require 'minitest/autorun'
+
+class ExteriorWallThermalPropertiesPercentChange_Test < Minitest::Test
+  def test_ExteriorWallThermalPropertiesPercentChange
+    # create an instance of the measure
+    measure = ExteriorWallThermalPropertiesPercentChange.new
+
+    # create an instance of a runner
+    runner = OpenStudio::Measure::OSRunner.new(OpenStudio::WorkflowJSON.new)
+
+    # load the test model
+    translator = OpenStudio::OSVersion::VersionTranslator.new
+    path = "#{File.dirname(__FILE__)}/example_model.osm"
+    model = translator.loadModel(path)
+    assert(!model.empty?)
+    model = model.get
+
+    # get arguments and test that they are what we are expecting
+    arguments = measure.arguments(model)
+    assert_equal(3, arguments.size)
+    assert_equal('r_value_perc_change', arguments[0].name)
+    assert_equal('solar_abs_perc_change', arguments[1].name)
+    assert_equal('thermal_mass_perc_change', arguments[2].name)
+
+    # set argument values to good values and run the measure on model with spaces
+    argument_map = OpenStudio::Measure::OSArgumentMap.new
+
+    r_value_perc_change = arguments[0].clone
+    assert(r_value_perc_change.setValue(1.2))
+    argument_map['r_value_perc_change'] = r_value_perc_change
+
+    solar_abs_perc_change = arguments[1].clone
+    assert(solar_abs_perc_change.setValue(1.4))
+    argument_map['solar_abs_perc_change'] = solar_abs_perc_change
+
+    thermal_mass_perc_change = arguments[2].clone
+    assert(thermal_mass_perc_change.setValue(0.3))
+    argument_map['thermal_mass_perc_change'] = thermal_mass_perc_change
+
+    measure.run(model, runner, argument_map)
+    result = runner.result
+    show_output(result)
+    assert(result.value.valueName == 'Success')
+  end
+end