+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 # *******************************************************************************
-# OpenStudio(R), Copyright (c) 2008-2020, Alliance for Sustainable Energy, LLC.
+# OpenStudio(R), Copyright (c) 2008-2021, Alliance for Sustainable Energy, LLC.
 # All rights reserved.
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions are met:
@@ -34,7 +34,7 @@
 # *******************************************************************************
 
 require 'bundler/setup'
-require 'openstudio/calibration_measures'
+require 'openstudio/calibration'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure